[CBR-226] Refactor: split Policy.hs into submodules
